Version 1.13
  * Fixed a problem with handling "left+right" vt fonts that was being conflated with italic/oblique.
  * On BSD console-fb-realizer now attempts to maintain the keyboard LED state.  Note that a quirk of VirtualBox, which employs subtle tricks like secret extra keypresses to keep host and guest keyboard lock states in synch, will make it seem like nothing is happening.
  * SoftTerm now supports RIS and DECSTR.
  * More cross-platform services: hostapd, kmod@, fail2ban, nagios
  * More BSD services: bthidd
  * Added an unload-when-stopped command.
  * nagios-check-service now has a --critical-if-below-min option.
  * service-dt-scanner no longer defaults to using input-activated log services.
  * Added a convert-fstab-services command.
  * SoftTerm now does VT52x sideways scrolling.

+++ source/service-control.cpp
@@ -92,12 +92,12 @@ service_control (
 		if (p.stopped()) throw EXIT_SUCCESS;
 	} catch (const popt::error & e) {
 		std::fprintf(stderr, "%s: FATAL: %s: %s\n", prog, e.arg, e.msg);
-		throw EXIT_USAGE;
+		throw static_cast<int>(EXIT_USAGE);
 	}
 
 	if (args.empty()) {
 		std::fprintf(stderr, "%s: FATAL: %s\n", prog, "Missing directory name(s).");
-		throw EXIT_USAGE;
+		throw static_cast<int>(EXIT_USAGE);
 	}
 
 	sigset_t original_signals;
